+++ packages/plan/src/Recovery/index.ts
@@ -1,3 +1,2 @@
 export * from './RecoveryPassPlan';
 export * from './RecoveryPlan';
-export * from './RecoverySpoilPlan';