+++ packages/schrodinger/src/Schrodinger.ts
@@ -2,16 +2,15 @@ import { Consumer, Serializable } from '@jamashita/anden-type';
 import { Alive } from './Alive';
 import { Contradiction } from './Contradiction';
 import { Dead } from './Dead';
-import { Detoxicated } from './Detoxicated';
 
 export type SchrodingerState = 'ALIVE' | 'CONTRADICTION' | 'DEAD' | 'STILL';
 
-export interface Schrodinger<A, D extends Error> extends Serializable {
-  get(): Detoxicated<A>;
+export interface Schrodinger<out A, out D extends Error> extends Serializable {
+  get(): Exclude<A, Error>;
 
   getState(): SchrodingerState;
 
-  ifAlive(consumer: Consumer<A>): void;
+  ifAlive(consumer: Consumer<Exclude<A, Error>>): void;
 
   ifContradiction(consumer: Consumer<unknown>): void;