+++ packages/unscharferelation/src/Plan/PresentPlan.ts
@@ -1,39 +1,38 @@
 import { Kind, UnaryFunction } from '@jamashita/anden-type';
-import { Matter, Nihil } from '@jamashita/genitore-heisenberg';
 import { MapPlan } from '@jamashita/genitore-plan';
 import { Epoque } from '../Epoque';
 import { isUnscharferelation, IUnscharferelation, UReturnType } from '../IUnscharferelation';
 
-export class PresentPlan<P, Q> implements MapPlan<Matter<P>> {
-  private readonly mapper: UnaryFunction<Matter<P>, UReturnType<Q>>;
+export class PresentPlan<in out P, out Q> implements MapPlan<Exclude<P, null | undefined | void>> {
+  private readonly mapper: UnaryFunction<Exclude<P, null | undefined | void>, UReturnType<Q>>;
   private readonly epoque: Epoque<Q>;
 
   public static of<P, Q>(
-    mapper: UnaryFunction<Matter<P>, UReturnType<Q>>,
+    mapper: UnaryFunction<Exclude<P, null | undefined | void>, UReturnType<Q>>,
     epoque: Epoque<Q>
   ): PresentPlan<P, Q> {
     return new PresentPlan(mapper, epoque);
   }
 
   protected constructor(
-    mapper: UnaryFunction<Matter<P>, UReturnType<Q>>,
+    mapper: UnaryFunction<Exclude<P, null | undefined | void>, UReturnType<Q>>,
     epoque: Epoque<Q>
   ) {
     this.mapper = mapper;
     this.epoque = epoque;
   }
 
-  private forOther(v: Nihil | Q): unknown {
+  private forOther(v: Q | null | undefined | void): unknown {
     if (Kind.isNone(v)) {
       return this.epoque.decline();
     }
 
-    return this.epoque.accept(v);
+    return this.epoque.accept(v as Exclude<Q, null | undefined | void>);
   }
 
   private forUnscharferelation(u: IUnscharferelation<Q>): unknown {
     return u.pass(
-      (v: Matter<Q>) => {
+      (v: Exclude<Q, null | undefined | void>) => {
         return this.epoque.accept(v);
       },
       () => {
@@ -45,16 +44,16 @@ export class PresentPlan<P, Q> implements MapPlan<Matter<P>> {
     );
   }
 
-  public onMap(value: Matter<P>): unknown {
+  public onMap(value: Exclude<P, null | undefined | void>): unknown {
     try {
       const mapped: UReturnType<Q> = this.mapper(value);
 
       if (isUnscharferelation(mapped)) {
         return this.forUnscharferelation(mapped);
       }
-      if (Kind.isPromiseLike<IUnscharferelation<Q> | Nihil | Q>(mapped)) {
+      if (Kind.isPromiseLike<IUnscharferelation<Q> | Q | null | undefined | void>(mapped)) {
         return mapped.then(
-          (v: IUnscharferelation<Q> | Nihil | Q) => {
+          (v: IUnscharferelation<Q> | Q | null | undefined | void) => {
             if (isUnscharferelation(v)) {
               return this.forUnscharferelation(v);
             }